requirements.txt
flask-mongoengine
Flask-Assets
bcrypt
bs4
